pyproject.toml
[tool.poetry]
name = "oac-to-fabric"
version = "3.0.0"
description = "Multi-agent migration framework: Oracle Analytics Cloud → Microsoft Fabric & Power BI"
authors = ["Migration Team"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.27"
pydantic = "^2.6"
pydantic-settings = "^2.2"
tenacity = "^8.2"
networkx = "^3.2"
lxml = "^5.1"
pyspark = "^3.5"
delta-spark = "^3.1"
azure-identity = "^1.15"
azure-keyvault-secrets = "^4.8"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.23"
pytest-cov = "^5.0"
ruff = "^0.3"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-p pytest_asyncio"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
